// src/apis/model/farmImageCoordinates.ts
/**
 * Generated by orval v6.14.3 🍺
 * Do not edit manually.
 * Farmer App API
 * OpenAPI spec version: v1
 */

export interface FarmImageCoordinates {
  fieldID?: string | null;
  imageUrl?: string | null;
  centerLat?: string | null;
  centerLatLarge?: string | null;
  centerLong?: string | null;
  centerLongLarge?: string | null;
  fieldLatLen?: string | null;
  fieldLatLenLarge?: string | null;
  fieldLongLen?: string | null;
  fieldLongLenLarge?: string | null;
  fieldMaxLat?: string | null;
  fieldMaxLong?: string | null;
  fieldMinLat?: string | null;
  fieldMinLong?: string | null;
}