// src/apis/model/responseDTO.ts
/**
 * Generated by orval v6.14.3 🍺
 * Do not edit manually.
 * Farmer App API
 * OpenAPI spec version: v1
 */
import type { ResponseDTOResult } from './responseDTOResult';

export interface ResponseDTO {
  isSuccess?: boolean;
  result?: ResponseDTOResult;
  displayMessage?: string | null;
  errorMessages?: string[] | null;
}